refactor(web): break TreeList component into files
