+++ packages/web/src/components/list/TreeList/index.js
@@ -30,17 +30,15 @@ import {
 	updateQuery as updateQueryAction,
 } from '@appbaseio/reactivecore/lib/actions/query';
 import { getInternalComponentID } from '@appbaseio/reactivecore/lib/utils/transform';
-import PreferencesConsumer from '../basic/PreferencesConsumer';
-import ComponentWrapper from '../basic/ComponentWrapper';
+import PreferencesConsumer from '../../basic/PreferencesConsumer';
+import ComponentWrapper from '../../basic/ComponentWrapper';
 
-import Container from '../../styles/Container';
-import Button from '../../styles/Button';
-import { HierarchicalMenuList, HierarchicalMenuListItem } from '../../styles/TreeList';
-import { connect } from '../../utils';
-import Input from '../../styles/Input';
+import Container from '../../../styles/Container';
+import { connect } from '../../../utils';
+import Input from '../../../styles/Input';
 
-import Title from '../../styles/Title';
-import { Checkbox, Radio } from '../../styles/FormControlList';
+import Title from '../../../styles/Title';
+import HierarchicalMenuComponent from './HierarchicalMenuComponent';
 
 const useConstructor = (callBack = () => {}) => {
 	const [hasBeenCalled, setHasBeenCalled] = useState(false);
@@ -64,253 +62,6 @@ const transformValueIntoLocalState = (valueArray) => {
 		valueToSet = newSelectedValues;
 	}
 	return valueToSet;
-};
-
-function HierarchicalMenuComponent({
-	listArray,
-	parentPath = '',
-	isExpanded = false,
-	listItemProps,
-}) {
-	if (!Array.isArray(listArray) || listArray.length === 0) {
-		return null;
-	}
-
-	return (
-		<HierarchicalMenuList className={`${isExpanded ? '--open' : ''}`} isSelected={isExpanded}>
-			{listArray.map(listItem => (
-				<HierarchicalMenuListItemComponent
-					key={`${parentPath}__${JSON.stringify(listItem)}`}
-					parentPath={parentPath}
-					listItem={listItem}
-					{...listItemProps}
-				/>
-			))}
-		</HierarchicalMenuList>
-	);
-}
-
-HierarchicalMenuComponent.propTypes = {
-	listArray: PropTypes.arrayOf(types.rawData), // array of objects
-	parentPath: types.string,
-	isExpanded: types.bool,
-	listItemProps: types.rawData,
-};
-const HierarchicalMenuListItemComponent = ({
-	selectedValues,
-	mode,
-	searchTerm,
-	listItem,
-	parentPath,
-	showLine,
-	renderItem,
-	handleListItemClick,
-	showCheckbox,
-	innerClass,
-	showRadio,
-	renderIcon,
-	showCount,
-	showSwitcherIcon,
-	switcherIcon,
-}) => {
-	if (!(listItem instanceof Object) || Object.keys(listItem).length === 0) {
-		return null;
-	}
-	const listItemLabel = listItem.key;
-	const listItemCount = listItem.count;
-	const isLeafNode = !(Array.isArray(listItem.list) && listItem.list.length > 0);
-
-	let newParentPath = listItemLabel;
-	if (parentPath) {
-		newParentPath = `${parentPath}.${listItemLabel}`;
-	}
-	let isSelected = false;
-	if (mode === 'single') {
-		if (recLookup(selectedValues, newParentPath) === true) {
-			isSelected = true;
-		}
-	} else {
-		isSelected = !!recLookup(selectedValues, newParentPath);
-	}
-
-	const [isExpanded, setIsExpanded] = useState(!!recLookup(selectedValues, newParentPath));
-
-	useEffect(() => {
-		setIsExpanded(!!recLookup(selectedValues, newParentPath));
-	}, [selectedValues]);
-
-	useEffect(() => {
-		if (listItem.initiallyExpanded) {
-			setIsExpanded(listItem.initiallyExpanded);
-		}
-	}, [listItem.initiallyExpanded]);
-
-	const renderSwitcherIcon = (isExpandedProp) => {
-		if (showSwitcherIcon === false) {
-			return null;
-		}
-		if (typeof switcherIcon === 'function') {
-			return switcherIcon(isExpandedProp);
-		}
-
-		return (
-			/* eslint-disable jsx-a11y/click-events-have-key-events
-			, jsx-a11y/no-static-element-interactions,jsx-a11y/no-noninteractive-tabindex */
-			<span
-				tabIndex="0"
-				onClick={(e) => {
-					e.stopPropagation();
-					setIsExpanded(!isExpanded);
-				}}
-				className="--switcher-icon"
-			>
-				&#10148;
-			</span>
-		);
-	};
-	return (
-		<HierarchicalMenuListItem
-			className={`${isSelected ? '-selected-item' : ''} ${
-				isExpanded ? '-expanded-item' : ''
-			}`}
-			key={newParentPath}
-			showLine={showLine}
-		>
-			<Button
-				isLinkType
-				onClick={() => {
-					handleListItemClick(listItemLabel, parentPath);
-				}}
-			>
-				{typeof renderItem === 'function' ? (
-					renderItem(listItemLabel, listItemCount, isSelected)
-				) : (
-					<React.Fragment>
-						{!isLeafNode && renderSwitcherIcon(isSelected)}
-						{/* eslint-disable jsx-a11y/click-events-have-key-events */}
-						{/* eslint-disable jsx-a11y/no-noninteractive-element-interactions */}
-						{mode === 'multiple' && showCheckbox && (
-							<React.Fragment>
-								<Checkbox
-									className={getClassName(innerClass, 'checkbox') || null}
-									checked={isSelected}
-									id={`${listItemLabel}-checkbox-${newParentPath}`}
-									name={`${listItemLabel}-checkbox-${newParentPath}`}
-									show
-									readOnly
-								/>
-								<label
-									style={{
-										width: '26px',
-										marginTop: 0,
-										marginBottom: 0,
-										marginRight: '-9px',
-										left: '-3px',
-									}}
-									htmlFor={`${listItemLabel}-checkbox-${newParentPath}`}
-									onClick={(e) => {
-										e.stopPropagation();
-									}}
-								/>
-							</React.Fragment>
-						)}
-						{mode === 'single' && showRadio && (
-							<React.Fragment>
-								<Radio
-									checked={isSelected}
-									className={getClassName(innerClass, 'radio') || null}
-									id={`${listItemLabel}-radio-${newParentPath}`}
-									name={`${listItemLabel}-radio-${newParentPath}`}
-									show
-									readOnly
-								/>
-
-								<label
-									style={{
-										width: '26px',
-										marginTop: 0,
-										marginBottom: 0,
-										marginRight: '-9px',
-										left: '-3px',
-									}}
-									htmlFor={`${listItemLabel}-radio-${newParentPath}`}
-									onClick={(e) => {
-										e.stopPropagation();
-									}}
-								/>
-							</React.Fragment>
-						)}{' '}
-						{/* eslint-enable jsx-a11y/click-events-have-key-events */}
-						{/* eslint-enable jsx-a11y/no-noninteractive-element-interactions */}
-						{renderIcon(isLeafNode)}
-						<div className="--list-item-label-count-wrapper">
-							<span
-								className={`--list-item-label ${
-									getClassName(innerClass, 'label') || ''
-								}`}
-							>
-								{listItemLabel}
-							</span>
-							{showCount && (
-								<span
-									className={`--list-item-count ${
-										getClassName(innerClass, 'count') || ''
-									}`}
-								>
-									{listItemCount}
-								</span>
-							)}
-						</div>
-					</React.Fragment>
-				)}
-			</Button>
-			{isLeafNode === false && (
-				<div className="--list-child">
-					{/* eslint-disable-next-line no-use-before-define */}
-					<HierarchicalMenuComponent
-						key={`${newParentPath}-${listItemLabel}-${listItemCount}`}
-						listArray={listItem.list}
-						parentPath={newParentPath}
-						isExpanded={isExpanded}
-						listItemProps={{
-							mode,
-							selectedValues,
-							searchTerm,
-							showLine,
-							renderItem,
-							handleListItemClick,
-							renderSwitcherIcon,
-							showCheckbox,
-							innerClass,
-							showRadio,
-							renderIcon,
-							showCount,
-							showSwitcherIcon,
-							switcherIcon,
-						}}
-					/>
-				</div>
-			)}
-		</HierarchicalMenuListItem>
-	);
-};
-HierarchicalMenuListItemComponent.propTypes = {
-	parentPath: types.string,
-	selectedValues: types.rawData,
-	mode: types.string,
-	searchTerm: types.string,
-	listItem: types.rawData,
-	showLine: types.bool,
-	renderItem: types.func,
-	handleListItemClick: types.func,
-	renderSwitcherIcon: types.func,
-	showCheckbox: PropTypes.bool,
-	innerClass: types.style,
-	showRadio: PropTypes.bool,
-	renderIcon: types.func,
-	showCount: PropTypes.bool,
-	showSwitcherIcon: types.bool,
-	switcherIcon: types.children,
 };
 
 const TreeList = (props) => {
@@ -344,7 +95,6 @@ const TreeList = (props) => {
 		switcherIcon,
 	} = props;
 	const hasMounted = useRef();
-
 	const [searchTerm, setSearchTerm] = useState('');
 	const [selectedValues, setSelectedValues] = useState({});
 
@@ -359,11 +109,11 @@ const TreeList = (props) => {
 			if (parentPath) {
 				newParentPath = `${parentPath}.${ele.key}`;
 			}
-			const keyHasSearchTerm
-				= replaceDiacritics(ele.key)
+			const keyHasSearchTerm =
+				replaceDiacritics(ele.key)
 					.toLowerCase()
-					.includes(replaceDiacritics(searchTerm).toLowerCase())
-				|| recLookup(selectedValues, newParentPath);
+					.includes(replaceDiacritics(searchTerm).toLowerCase()) ||
+				recLookup(selectedValues, newParentPath);
 
 			if (isLeafItem && keyHasSearchTerm) {
 				result.push({
@@ -424,7 +174,7 @@ const TreeList = (props) => {
 
 		if (value) {
 			// adds a sub-query with must as an array of objects for each term/value
-			const queryArray = value.map(item => ({
+			const queryArray = value.map((item) => ({
 				[type]: {
 					[props.dataField]: item,
 				},
@@ -481,8 +231,8 @@ const TreeList = (props) => {
 	}
 
 	const setValue = (value, hasMountedParam = hasMounted.current) => {
-		const finalValues
-			= Array.isArray(value) === false
+		const finalValues =
+			Array.isArray(value) === false
 				? transformTreeListLocalStateIntoQueryComptaibleFormat(value)
 				: value;
 		const performUpdate = () => {
@@ -529,8 +279,8 @@ const TreeList = (props) => {
 
 	useEffect(() => {
 		if (hasMounted.current) {
-			const valueArray
-				= transformTreeListLocalStateIntoQueryComptaibleFormat(selectedValues) || [];
+			const valueArray =
+				transformTreeListLocalStateIntoQueryComptaibleFormat(selectedValues) || [];
 			updateQuery(valueArray);
 		}
 	}, [props.customQuery]);
@@ -543,8 +293,8 @@ const TreeList = (props) => {
 
 	useEffect(() => {
 		if (hasMounted.current) {
-			const valueArray
-				= transformTreeListLocalStateIntoQueryComptaibleFormat(selectedValues) || [];
+			const valueArray =
+				transformTreeListLocalStateIntoQueryComptaibleFormat(selectedValues) || [];
 			updateQueryOptions();
 			updateQuery(valueArray);
 		}
@@ -613,11 +363,11 @@ const TreeList = (props) => {
 		return null;
 	};
 
-	const sanitizeObject = obj =>
+	const sanitizeObject = (obj) =>
 		JSON.parse(
 			JSON.stringify(obj, (key, value) =>
 				// eslint-disable-next-line eqeqeq
-				(value === null || value == {} || value === false ? undefined : value),
+				value === null || value == {} || value === false ? undefined : value,
 			),
 		);
 
@@ -639,17 +389,15 @@ const TreeList = (props) => {
 		if (props.value === undefined) {
 			setValue(newSelectedValues);
 		} else if (props.onChange) {
-			const valueToSet
-				= transformTreeListLocalStateIntoQueryComptaibleFormat(newSelectedValues);
+			const valueToSet =
+				transformTreeListLocalStateIntoQueryComptaibleFormat(newSelectedValues);
 
 			props.onChange(valueToSet);
 		}
 	};
 
 	const renderIcon = (isLeafNode) => {
-		const {
-			showIcon, showLeafIcon, icon, leafIcon,
-		} = props;
+		const { showIcon, showLeafIcon, icon, leafIcon } = props;
 
 		if (isLeafNode) {
 			if (!showLeafIcon) return null;
@@ -823,9 +571,9 @@ TreeList.defaultProps = {
 
 const mapStateToProps = (state, props) => ({
 	selectedValue:
-		(state.selectedValues[props.componentId]
-			&& state.selectedValues[props.componentId].value)
-		|| null,
+		(state.selectedValues[props.componentId] &&
+			state.selectedValues[props.componentId].value) ||
+		null,
 	rawData: state.rawData[props.componentId] || {},
 	aggregationData: state.aggregations[props.componentId] || {},
 	themePreset: state.config.themePreset,
@@ -834,19 +582,19 @@ const mapStateToProps = (state, props) => ({
 	enableAppbase: state.config.enableAppbase,
 });
 
-const mapDispatchtoProps = dispatch => ({
+const mapDispatchtoProps = (dispatch) => ({
 	setQueryOptions: (component, props) => dispatch(setQueryOptionsAction(component, props)),
-	updateQuery: updateQueryObject => dispatch(updateQueryAction(updateQueryObject)),
+	updateQuery: (updateQueryObject) => dispatch(updateQueryAction(updateQueryObject)),
 });
 
 const ConnectedComponent = connect(
 	mapStateToProps,
 	mapDispatchtoProps,
-)(withTheme(props => <TreeList ref={props.myForwardedRef} {...props} />));
+)(withTheme((props) => <TreeList ref={props.myForwardedRef} {...props} />));
 
 const ForwardRefComponent = React.forwardRef((props, ref) => (
 	<PreferencesConsumer userProps={props}>
-		{preferenceProps => (
+		{(preferenceProps) => (
 			<ComponentWrapper
 				{...preferenceProps}
 				internalComponent
